integration project tp2

+++ Integration project/tp2/indexF.php
@@ -1,26 +1,61 @@
 <?php
 if ($statut == "succes") {
-  $req = "SELECT * from RatVol";
-  $result = $mysqli->query($req);
-  function printResult(){
-    global $result ; 
-    while ($row = $result->fetch_assoc()) {
-      echo "<tr>";
-      echo "<td>" . $row['NumRatV'] . "</td>";
-      echo "<td>" . $row['MatProf'] . "</td>";
-      echo "<td>" . $row['DateRat'] . "</td>";
-      echo "<td>" . $row['Seance'] . "</td>";
-      echo "<td>" . $row['Session'] . "</td>";
-      echo "<td>" . $row['Salle'] . "</td>";
-      echo "<td>" . $row['Jour'] . "</td>";
-      echo "<td>" . $row['CodeClasse'] . "</td>";
-      echo "<td>" . $row['CodeMatiere'] . "</td>";
-      echo "<td>" . $row['Etat'] . "</td>";
-      echo '<td><a style="color: blue;" href="editForm.php?id=' . $row['NumRatV'] . '">Edit</a> | <a style="color: red;"  href="delete.php?id=' . $row['NumRatV'] . '">Delete</a></td>';
-      echo "</tr>";
+  $NumRatV = $_POST["filterNumRatV"];
+  $MatProf = $_POST["filterMatProf"];
+  $DateRat = $_POST["filterDateRat"];
+  $Seance = $_POST["filterSeance"];
+  $Session = $_POST["filterSession"];
+  $Salle = $_POST["filterSalle"];
+  $Jour = $_POST["filterJour"];
+  $CodeClasse = $_POST["filterCodeClasse"];
+  $CodeMatiere = $_POST["filterCodeMatiere"];
+  $Etat = $_POST["filterEtat"];
+  $header = "addForm";
+
+
+  function printResult()
+  {
+    $filterR = 0;
+    $filter = [];
+    global $mysqli;
+    global $result, $MatProf, $DateRat, $Seance, $Session, $Salle, $Jour, $CodeClasse, $CodeMatiere, $NumRatV;
+    $table = [$NumRatV, $MatProf, $DateRat, $Seance, $Session, $Salle, $Jour, $CodeClasse, $CodeMatiere];
+
+    for ($i = 0; $i < count($table); $i++) {
+      if ($table[$i] !== "" && !is_null($table[$i])) {
+        $filter[] = $table[$i];
+        $filterR = 1;
+      }
     }
+
+    echo $filterR;
+
+    foreach ($filter as $value) {
+      echo $value . "<br>";
+    }
+
+    if ($filterR == 0) {
+      $req = "SELECT * from RatVol";
+      $result = $mysqli->query($req);
+      while ($row = $result->fetch_assoc()) {
+        echo "<tr>";
+        echo "<td>" . $row['NumRatV'] . "</td>";
+        echo "<td>" . $row['MatProf'] . "</td>";
+        echo "<td>" . $row['DateRat'] . "</td>";
+        echo "<td>" . $row['Seance'] . "</td>";
+        echo "<td>" . $row['Session'] . "</td>";
+        echo "<td>" . $row['Salle'] . "</td>";
+        echo "<td>" . $row['Jour'] . "</td>";
+        echo "<td>" . $row['CodeClasse'] . "</td>";
+        echo "<td>" . $row['CodeMatiere'] . "</td>";
+        echo "<td>" . $row['Etat'] . "</td>";
+        echo '<td><a style="color: blue;" href="editForm.php?id=' . $row['NumRatV'] . '">Edit</a> | <a style="color: red;"  href="delete.php?id=' . $row['NumRatV'] . '">Delete</a></td>';
+        echo "</tr>";
+      }
+    }
+    
   }
 } else {
   echo $statut;
 }
-?>
+
